Remove TODOs for read/write callbacks auth.
Turns out Bluetooth MAC addresses are frequently randomized, so it's not a good
idea to check auth solely based on MAC addresses. On top of that, checking auth
during read/write is probably too late, and it's better to check it during
initial connection. Instead, one should probably use BLESecurity or something
like that.